Fix types and a bit of refactoring
