+++ src/stories/Timeline.stories.tsx
@@ -1,10 +1,13 @@
+import Chance from 'chance';
 import dayjs from 'dayjs';
 import React from 'react';
 import { ComponentMeta, ComponentStory } from '@storybook/react';
 
+const chance = Chance();
+
 import mockUpdate from 'utils/testing/mocks/mockUpdate';
 import Timeline from 'components/Timeline';
-import { UPDATE_TYPES } from 'types/updates';
+import { UPDATE_TYPES, ZetkinUpdateJourneyMilestone } from 'types/updates';
 
 export default {
   argTypes: {
@@ -27,13 +30,19 @@ const addAssigneeUpdates = Array.from(Array(10).keys()).map(() =>
   })
 );
 
-const journeyMilestoneUpdates = Array.from(Array(10).keys()).map(() =>
-  mockUpdate(UPDATE_TYPES.JOURNEYINSTANCE_UPDATEMILESTONE, {
+const journeyMilestoneUpdates = Array.from(Array(10).keys()).map(() => {
+  const update = mockUpdate(UPDATE_TYPES.JOURNEYINSTANCE_UPDATEMILESTONE, {
     timestamp: dayjs()
       .subtract(Math.random() * 100, 'hours')
       .format(),
-  })
-);
+  }) as ZetkinUpdateJourneyMilestone;
+  update.details.milestone.title = chance.sentence({ words: 4 });
+  if (Math.random() > 0.5) {
+    const change = update.details.changes.completed;
+    update.details.changes.completed = { from: change.to, to: change.from };
+  }
+  return update;
+});
 
 const updates = addAssigneeUpdates
   .concat(journeyMilestoneUpdates)
